Module 3 Lab Activity Exercise Files
These are my Module 3 Lab Activity Exercise Files for the week. *(I wasn't quite clear on the Name and Instructor Greeting instructions so  I’ve also included a variation of the code for this program called “Name Instructor Greeting2.py - This program requests input of a username and the first name of the instructor. It is strict in who it responds to since the variables that are assigned to the objects are very specific. You can modify the variables with other specific names however, each statement must correspond with the specific names assigned. All other names will not be recognized by the program.)